Added folder segregation and errorcatches from RPi
